requirements.txt
opencv-python-headless~=4.5.2.52
numpy~=1.20.2
streamlit~=0.81.0
Pillow~=8.1.0